requirements.txt
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
google-ai-generativelanguage==0.6.2
google-auth==2.29.0
google-generativeai==0.5.0
idna==3.6
protobuf==4.25.3
python-dotenv==1.0.1
requests==2.31.0
soupsieve==2.5
urllib3==2.2.1